Required Interface

CommonRLInterface has a basic required interface for all environments consisting of the following functions documented below:

reset!

actions

observe

act!

terminated

CommonRLInterface.reset! — Function

reset!(env::AbstractEnv)

Reset env to its initial state and return nothing.

This is a required function that must be provided by every AbstractEnv.

CommonRLInterface.actions — Function

actions(env::AbstractEnv)

Return a collection of all of the actions available for AbstractEnv env. If the environment has multiple agents, this function should return the union of all of their actions.

This is a required function that must be provided by every AbstractEnv.

This function is a static property of the environment; the value it returns should not change based on the state.

actions(env::AbstractEnv, i::Integer)

Return a collection of all the actions available to player i.

This function is a static property of the environment; the value it returns should not change based on the state.

CommonRLInterface.observe — Function

observe(env::AbstractEnv)

Return an observation from the environment for the current player.

This is a required function that must be provided by every AbstractEnv.

CommonRLInterface.act! — Function

r = act!(env::AbstractEnv, a)

Take action a for the current player, advance AbstractEnv env forward one step, and return rewards for all players.

This is a required function that must be provided by every AbstractEnv.

CommonRLInterface.terminated — Function

terminated(env::AbstractEnv)

Determine whether an environment has finished executing.

If terminated(env) is true, no further actions should be taken and it is safe to assume that no further rewards will be received.